refactor(server): move canonical document runtime into package
