+++ packages/doc-editor/package.json
@@ -15,6 +15,7 @@
     "./schema/code-block-ext": "./src/schema/code-block-ext.ts",
     "./schema/frontmatter": "./src/schema/frontmatter.ts",
     "./schema/proof-marks": "./src/schema/proof-marks.ts",
+    "./utils/text-range": "./src/utils/text-range.ts",
     "./plugins/comments": "./src/plugins/comments.ts",
     "./plugins/marks": "./src/plugins/marks.ts",
     "./plugins/suggestions": "./src/plugins/suggestions.ts"
