+++ packages/doc-server/src/collab-shared.ts
@@ -5,6 +5,8 @@ export {
   applyCanonicalDocumentToCollab,
   applyCanonicalDocumentToCollabWithVerification,
   buildCollabSession,
+  detectPathologicalProjectionRepeat,
+  evaluateProjectionSafety,
   extractCollabTokenFromHeaders,
   getCanonicalReadableDocumentSync,
   getCollabRuntime,
@@ -24,6 +26,7 @@ export {
   loadCanonicalYDoc,
   logCollabSocketErrorWithSuppression,
   noteRecentCollabSessionLease,
+  registerCanonicalYDocPersistence,
   releaseRewriteLock,
   removeAgentPresenceFromLoadedCollab,
   stripEphemeralCollabSpans,
